Agregando funcionalidad en login y logout para escoger la sede a la que quiere estas relacionado para los usuarios sin loguear

+++ app/src/main/java/com/co/edu/cun/www1104379214/bienestarcun/SqliteBD/TaskExecuteSQLDelete.java
@@ -50,7 +50,8 @@ public class TaskExecuteSQLDelete extends AsyncTask<Void, Void, Integer> {
 
         Integer resultado = null;
 
-        resultado = db.DeleteBD();
+        if( db != null )
+            resultado = db.DeleteBD();
 
         return resultado;
